Automatically derive PR for github-mark-ready (#6457)
# Motivation

`scripts/github-mark-ready` lets you automatically mark a PR as "Ready
for review" once all the checks has passed.
Currently you need to pass the URL of which PR you want to mark as
ready.
With this PR you don't have to pass a URL if you want it to apply to the
PR of your current branch.

# Changes

1. Determine `$PR` and `$REPO` automatically from the current branch
when no argument is passed.

# Tests

Manually

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary