feat(tokens): handle failed balances for out-of-cycles canisters (#6469)
# Motivation

We don't want to consider the balances of failed accounts for tokens
during loading. This is to prevent displaying incorrect information to
users, which might lead them to believe that the token will load
eventually.

This PR utilizes an out-of-cycle store to return a failed account
balance when loading user tokens.

Note: On the Tokens page, the token is now displayed with a warning
message and the `canisterId`. PR #6470 addresses the use of the token
name if available; otherwise, it defaults to the canister ID.

Before:


https://github.com/user-attachments/assets/778480a7-87a0-417d-acfe-efbf63e03797

After:


https://github.com/user-attachments/assets/994e9a8c-24b0-4881-927d-e30cc72a2ae6

[NNS1-3619](https://dfinity.atlassian.net/browse/NNS1-3619)

# Changes

- `convertToUserTokenData` to return failed balance for ledger canisters
out of cycles

# Tests

- Adds unit test

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary

Prev. PR: #6466 

[NNS1-3619]:
https://dfinity.atlassian.net/browse/NNS1-3619?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ