chore(config): centralize node version management in config.json (#6479)
# Motivation

Following up on #6431, we aim to establish a single source of truth for
the node version in the project. To remain consistent with other
versions, we will use the one defined in `config.json`.

Note: Some GitHub actions currently use `.nvmrc` to determine the node
version. These will be changed in a follow-up PR.

# Changes

- Removed `NODE_VERSION` from the `Dockerfile` as it was not used.  
- Added a script to populate `.nvmrc` from `config.json` and exposed it
as an npm script.
- Removes `.nvmrc` from tracked files since it can now be generated
locally.
- Updated the `README` with local development instructions and
information about `.nvmrc`.

# Tests

- GitHub actions should run as before.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary