Make ProjectDetail.spec.ts more robust (#6499)
# Motivation

During the `"should participate without user interaction if there is an
open ticket."` test in `ProjectDetail.spec.ts`, the
`SaleInProgressModal` becomes visible and then invisible, after which
the commitment should have appeared. Meanwhile `querySnsSwapCommitment`
is first called twice and then once more.

The test expects to catch all of this without making sure that the
process is paused at any point in time. This is very brittle because it
depends on the framework waiting asynchronously at various steps.

With the Svelte 5 update this breaks, unsurprisingly.

The solution is to make sure that the test is blocked in states that we
want to expect on.

# Changes

1. Return unresolved promises from `querySnsSwapCommitment` and only
resolve them after verifying the state of the component.
2. Verify that the 2 calls are for "query" and "update" instead of just
claiming it in a comment.

# Tests

1. Still passes.
2. Now also passes in the [Svelte 5
branch](https://github.com/dfinity/nns-dapp/pull/6489).

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary