feat(ui): apply dynamic monetary value formatter for total values (#6512)
# Motivation

The nns-dapp displays monetary values in several sections: the
portfolio, tokens, and projects page. These values can sometimes be
large numbers with excessive decimal detail.

This second PR uses the utility introduced in #6507 for the total values
shown in the portfolio, tokens, and projects page.

A third PR will update the TopHeld cards.

[NNS1-3604](https://dfinity.atlassian.net/browse/NNS1-3604)

# Changes

- Changes the format utility used in `UsdValueHeadless`.

# Tests

- Updates tests to reflect the new values

# Todos

- [x] Add entry to changelog (if necessary).

[NNS1-3604]:
https://dfinity.atlassian.net/browse/NNS1-3604?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ