feat(portfolio): add participation amount formatter util (#6531)
# Motivation

For the upcoming changes to the Portfolio page, we need to display
certain values of the Open Proposals, such as the minimum and maximum
commitment. We want to present these with special formatting. Please see
the attached screenshot.

This PR introduces a new utility for the Portfolio page to format
numbers accordingly.

<img width="681" alt="Screenshot 2025-03-04 at 15 56 51"
src="https://github.com/user-attachments/assets/14b4cc79-b730-43bb-a861-60af3fedf1e1"
/>

# Changes

- Adds a new utility to format participation on the portfolio page.

# Tests

- New unit tests.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.