Redesign tokens table sorting (#6536)
# Motivation

Currently, on the tokens page, there is no sorting option available on
mobile. The spot where the open-sorting-modal button is shown in other
tables is occupied by the settings button, which allows hiding zero
tokens (controlled by the disableMobileSorting prop).

With this PR, we add the ability to access sorting from the settings
popover in the tokens table. Replacing sorting from the modal with
sorting from the popover in the neurons and projects tables will be
handled in a separate PR.

# Changes

- Remove disableMobileSorting support. It was used only in the tokens
table.
- Display open-sort-modal button only when there is no setting button.
Currently the setting button is shown only in tokens table.
- Always display the ResponsiveTableSortControl in the settings popover.
The isSortingEnabled flag is ignored because all current tables have
sortable columns, so I find it redundant.

# Tests

- Updated.
- Tested manually.

Demo:
https://qsgjb-riaaa-aaaaa-aaaga-cai.mstr-ingress.devenv.dfinity.network/tokens/

# Todos

- [ ] Add entry to changelog (if necessary).