feat(portfolio): add minimum commitment percentage calculator (#6537)
# Motivation

For the upcoming changes to the Portfolio page, we need a percentage of
a project's minimum ICP participation for the project to be successful.

This PR introduces a new utility to extract values from a
`FullProjectCommitmentSplit` and perform the calculation.

<img width="202" alt="Screenshot 2025-03-05 at 13 16 07"
src="https://github.com/user-attachments/assets/f1b25f2a-5ca1-4fcc-8d16-cd9c449a9b8e"
/>

# Changes

- New portfolio utility to calculate the percentage of the minimum
commitment that has already been raised.

# Tests

- Added unit tests

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary