chore(ci): update rustup to 1.28.1 and remove toolchain install workaround (#6548)
# Motivation

The Reproducible Docker Builds
[failed](https://github.com/dfinity/nns-dapp/actions/runs/13643019276)due
a new version of
[rustup](https://blog.rust-lang.org/2025/03/02/Rustup-1.28.0.html) due
to a new version of rustup.

#6529 introduced this new version but also required additional changes
because the auto-installation of the toolchain is no longer done
automatically.

This change appears to be unintended, as it has been addressed in the
latest
[patch](https://github.com/rust-lang/rustup/blob/master/CHANGELOG.md#1281---2025-03-05)
of rustup.

This PR updates the `rustup` version to the fixed release and cleans the
unnecessary code.

# Changes

- Updates the `rustup` version to the latest patch.
- Removes unnecessary code that was used to install the toolchain with
`rustup`.

# Tests

- The pipeline should pass.
- Manually trigger the failed job for this branch:
https://github.com/dfinity/nns-dapp/actions/runs/13722208821

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.