test: import logo (#6559)
# Motivation

In the Svelte v5 branch, a test for a logo is failing because, instead
of referencing the SVG, the content is inlined. I assume you don't want
to test whether the SVG is inlined or not, but rather test if the
expected logo is used. Therefore, this PR imports the logo and uses the
result for comparison purposes.

Signed-off-by: David Dal Busco <david.dalbusco@dfinity.org>