fix(errors): update cycles detection to use error codes (#6564)
# Motivation

#6465 introduced a method for tracking specific errors based on the
error message payload. This functionality broke after the following
change: https://github.com/dfinity/ic/pull/3465.

The
[recommendation](https://dfinity.slack.com/archives/CGA566TPV/p1741614348749269?thread_ts=1741613964.837469&cid=CGA566TPV)
is to utilize the fields instead of parsing the message.

Please note that we need to use type guards to retrieve information
about the errors, as it is not currently propagated to the client app.

# Changes

- Refactor `isCanisterOutOfCyclesError` to narrow down error's type and
utilize the `error_code` information.

# Tests

- Updated tests based on new implementation.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary