fix(ci): allow unencrypted identities for app environment deployment (#6575)
# Motivation

We have multiple environments available for testing new features before
they are released to `mainnet`. One environment is `beta`, currently
controlled through Orbit, while the other is `app`. The `beta`
environment points to the main branch and is managed by Orbit, whereas
the `app` environment can be triggered as a GitHub action for any
branch.

This PR addresses an issue with the GitHub job. Starting with dfx
0.25.0, canister calls to mainnet using unencrypted identities are no
longer permitted.

# Changes

- Disables the warning for the `app` deployment

# Tests

- Successfully ran the
[job](https://github.com/dfinity/nns-dapp/actions/runs/13833020107)

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary