feat(neurons): collapse SNS voting power details into expandable section (#6585)
# Motivation

Following up on the work done in #6138, we also want to hide details
about voting power behind an accordion-like component for SNS neurons.
This way, users must proactively click to expand and see the details.
This change will simplify the complexity of the UI.

Before:
![Screenshot 2025-03-16 at 10 08
38](https://github.com/user-attachments/assets/e980efa3-d348-4ab9-9d2a-a1b27ab239de)

After:


https://github.com/user-attachments/assets/806512ef-9044-49c7-8f2c-e28b3be6cd63

[NNS1-3660](https://dfinity.atlassian.net/browse/NNS1-3660)

# Changes

Use `KeyValuePairInfo` to conceal the voting calculation logic.

# Tests

- Should pass as before

# Todos

- [x] Add entry to changelog (if necessary).


[NNS1-3660]:
https://dfinity.atlassian.net/browse/NNS1-3660?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ