Sns topics (#6582)
# Motivation

To efficiently access the list of all topics for each SNS, the
sns-aggregator provides a new field, topics, which contains the
SnsGov.list_topics endpoint response as JSON data (the
[pr](https://github.com/dfinity/nns-dapp/pull/6547)).

To make it compatible with ic-types, we need to convert them back. Here,
we provide new conversion utility functions to handle this.

# Changes

- Add DTO types.
- New UnknownTopic type in case the aggregator is outdated.
- New conversion utils 
  - convertDtoToTopic
  - convertDtoTopicInfo
  - convertDtoToListTopicsResponse

# Tests

- Updated.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.