feat(transactions): display fee amount in USD (#6598)
# Motivation

Following up on #6597, we want to display the USD value for
transactions. In this PR, we are adding a new section to the
`TransactionFormFee` to show the fee in dollars.
We use the same data source for the token prices as in the rest of the
application, icp-swap.

<img width="640" alt="Screenshot 2025-03-19 at 11 34 25"
src="https://github.com/user-attachments/assets/7985d053-bccf-4dba-9d23-fc957391224b"
/>

[NNS1-3671](https://dfinity.atlassian.net/browse/NNS1-3671)

# Changes

- Displays the USD value of the transaction fee

# Tests

- Adds unit tests for the component

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet.

[NNS1-3671]:
https://dfinity.atlassian.net/browse/NNS1-3671?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ