feat(sns_aggregator): Collect data from the new `SnsGov.list_topics` endpoint (#6547)
# Motivation

We want the NNS dapp to have efficient access to the list of all topics
(and proposals registered under each topic) for each SNS. This is
provided by the new `SnsGov.list_topics` endpoint, so this PR extends
the SNS Aggregator s.t. the new endpoint is being periodically called
and the respective data is aggregated.

# Changes

* Added `SlowSnsData.topics` and `UpstreamData.topics`.

# Tests

<!-- Please provide any information or screenshots about the tests that
have been done -->

# Todos

- [v] Add entry to changelog.
- [ ] Manual e2e testing.