Removing anything related to SnsAction (#6570)
# Motivation

There was an outage when the `action` field from the `ProposalData`
(from SNS Governance) cannot be parsed. There is an agent-js bug which
makes the candid parsing fail. However, even if the candid parsing bug
is fixed, it was unclear whether the NNS Dapp would be able to show the
SNS proposal correctly if it's unaware of the new action variant (and
`action` is decoded to null). This PR tries to prove that it indeed can,
by removing any access to the `action` field.

# Changes

- Remove `SnsProposalDataMap::actionData`
- Remove any access to `actionData`, which all happen to be dead code
- Stop converting `action` in `mapProposalInfo`

Note that after those changes, the only place `SnsAction` (from
`@dfinity/sns`) is used is `frontend/src/lib/api/sns-dummy.api.ts`,
which is for SUBMITTING proposals, and those will still work if the
field `ProposalData::action` (in the list_proposals RESPONSE) does not
exist. However, the PR does not definitively prove that NNS Dapp still
works if `ProposalData::action` does not exist, since `.action` can
still be called somewhere and I cannot find a reliable way to finding
them

# Tests

N/A

# Todos

- [ ] Add entry to changelog (if necessary).