sns topics: Topic item component (#6624)
# Motivation

Migrating from following by type to following by topic to improve neuron
management.

In the new “Follow” modal, all topics will be displayed with the ability
to select them for following in the next step.
This PR adds the component that will be used in the topic list. The full
component structure can be found in the draft pr -
https://github.com/dfinity/nns-dapp/pull/6610.

Currently **DEMO NOT WORKING** because of the missing aggregator update.

[Demo](https://qsgjb-riaaa-aaaaa-aaaga-cai.mstr-ingress.devenv.dfinity.network/neuron/?u=7tjcv-pp777-77776-qaaaa-cai&neuron=a12652a79755ee2158248250b9038a35a87adbd53674d789b318bf024e1d2989)
1. `__featureFlags.ENABLE_SNS_TOPICS.overrideWith(true)`
2. Click `Follow Neurons` button on the sns neuron detail page.



https://github.com/user-attachments/assets/1de40535-f9d2-43c5-b87b-07c5add8ef5c


# Changes

- New FollowSnsNeuronsByTopicItem component

# Tests

- Added.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.