style(neurons): relocate following section above fold (#6629)
# Motivation

Similar to #6628, we want to enhance the visibility of the following
feature on the Neuron's page. In this PR, we move the Following section
above the fold so that users can see it as soon as they land on the
page.

[NNS1-3662](https://dfinity.atlassian.net/browse/NNS1-3662)

# Changes

- Move the location of the following card to the top of the page, right
below the Voting Power.

# Tests

<!-- Please provide any information or screenshots about the tests that
have been done -->

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3662]:
https://dfinity.atlassian.net/browse/NNS1-3662?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ