style(portfolio): optimize total assets card layout  (#6693)
# Motivation

We want to optimize the display of the `TotalAssetsCard` based on the
presence of StackedCards. When there are no Cards, we aim to maximize
the horizontal space. However, when Cards are present, we want to
maximize the vertical space.

[NNS1-3641](https://dfinity.atlassian.net/browse/NNS1-3641)

With `StackedCards`:

<img width="1499" alt="Screenshot 2025-04-14 at 09 23 43"
src="https://github.com/user-attachments/assets/d160f091-0d2e-4196-be8c-1bd40bf7c3fd"
/>

Without `StackedCards`:

<img width="1512" alt="Screenshot 2025-04-14 at 09 24 44"
src="https://github.com/user-attachments/assets/2784b468-feba-461d-9cea-f38195d35f8c"
/>

# Changes

- Add conditional CSS styles when no cards are present.

# Tests

- Adds unit tests to both the `TotalAssetsCard` and the `Portfolio` page
based on the presence of SNS cards or not.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.

[NNS1-3641]:
https://dfinity.atlassian.net/browse/NNS1-3641?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ