chore(scripts): simplify beta release verification steps (#6685)
# Motivation

The `scripts/nns-dapp/release-beta` creates an Orbit proposal to release
the nns-dapp to Beta. Voters should ideally review the proposal before
casting their votes. Currently, the instructions for reviewing the
proposal are as follows:
```
1. git fetch
2. git checkout $COMMIT
3. git merge-base --is-ancestor HEAD origin/main && echo OK || echo \"Commit is not on main branch!\"
4. ./scripts/docker-build
5. sha256sum nns-dapp.wasm.gz"
```

This is very helpful for understanding what is happening, but it is a
bit cumbersome to operate since one has to copy and paste each line.

# Changes

- Adds a one-liner with instructions to verify the proposal.

# Tests

- Tested locally.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.